Fix bugs and optimize code

+++ libmanuels.py
@@ -1,12 +1,28 @@
 import os
+import sys
 import signal
 import time
 from selenium import webdriver
-from selenium.webdriver.chrome.service import Service as ChromeService
-from selenium.webdriver.chrome.options import Options as ChromeOptions
+from selenium.webdriver.edge.options import Options
 from pdfu import generate_pdf, merge_pdfs
 from colorama import Fore, Style, init
-from webdriver_manager.chrome import ChromeDriverManager
+
+if not os.path.exists("manuels"):
+    os.makedirs("manuels")
+
+def get_arg(field):
+    if len(sys.argv) == 1:
+        return None
+    match field:
+        case "manid":
+            if "--id" in sys.argv:
+                return sys.argv[sys.argv.index("--id") + 1]
+        case "edition":
+            if "--edition" in sys.argv:
+                return sys.argv[sys.argv.index("--edition") + 1]
+        case "delete":
+            return "--no-delete" in sys.argv
+
 
 editor_dict = {
     1: "Magnard",
@@ -17,26 +33,18 @@ editor_dict = {
 init(autoreset=True)
 
 start_time = time.time()
-chrome_options = ChromeOptions()
-chrome_options.add_argument('--headless')
-chrome_options.add_argument('--disable-gpu')
-chrome_options.add_argument("--log-level=3")
+options = Options()
+options.add_argument('--headless')
+options.add_argument('--disable-gpu')
+options.add_argument("--log-level=3")
 
-chrome_install = ChromeDriverManager().install()
-
-folder = os.path.dirname(chrome_install)
-if os.name == 'nt':
-    chromedriver_path = os.path.join(os.path.dirname(chrome_install), "chromedriver.exe")
-else:
-    chromedriver_path = os.path.join(os.path.dirname(chrome_install), "chromedriver")
     
-driver = webdriver.Chrome(
-    service = ChromeService(chromedriver_path),
-    options=chrome_options
+driver = webdriver.Edge(
+    options=options
 )
 
 os.system('cls' if os.name == 'nt' else 'clear')
-print(f"""{Fore.RED}
+print(fr"""{Fore.RED}
 
 $$\      $$\                                         $$\       $$$$$$$\                                                        
 $$$\    $$$ |                                        $$ |      $$  __$$\                                                       
@@ -52,7 +60,7 @@ $$ | \_/ $$ |\$$$$$$$ |$$ |  $$ |\$$$$$$  |\$$$$$$$ |$$ |      $$$$$$$  |\$$$$$$
 
 {Style.RESET_ALL}
 {Fore.YELLOW}libmanuels.Fr edititon. {Style.RESET_ALL} To find books to dump, use this https://demo.edulib.fr/bibliotheque and check the book id in the url. Ex: 9782210783133
-By {Fore.BLUE}Ulysse2211{Style.RESET_ALL}, thanks to the contribution of {Fore.RED}miickboy{Style.RESET_ALL} and {Fore.RED}noh.am{Style.RESET_ALL}
+By {Fore.BLUE}Ulysse2211{Style.RESET_ALL}, thanks to the contribution of {Fore.RED}miickboy{Style.RESET_ALL}, {Fore.RED}noh.am{Style.RESET_ALL} and {Fore.RED}codeuriii{Style.RESET_ALL}
       """)
 
 print("""
@@ -64,18 +72,19 @@ Editor choices:
 4) La Librairie des écoles
 5) Other
       """)
-editor = input("Choose a number for the name for the editor of the book you want to dump. ")
+editor = get_arg("edition") or input("Choose a number for the name for the editor of the book you want to dump. ")
 if editor == "5":
     editor = input("\nEnter the name of the editor you want to dump (case-sensitive): ")
 else:
-    try:
-        editor = editor_dict[int(editor)]
-    except (KeyError, ValueError):
-        os.system('cls' if os.name == 'nt' else 'clear')
-        print(Fore.RED + "Invalid editor number. Please choose a number between 1 and 5." + Style.RESET_ALL)
-        exit()
+    if not editor in editor_dict.values():
+        try:
+            editor = editor_dict[int(editor)]
+        except (KeyError, ValueError):
+            os.system('cls' if os.name == 'nt' else 'clear')
+            print(Fore.RED + "Invalid editor number. Please choose a number between 1 and 5." + Style.RESET_ALL)
+            exit()
 
-manid = input(Fore.CYAN + "\nEnter the ID of the manual: " + Style.RESET_ALL)
+manid = get_arg("manid") or input(Fore.CYAN + "\nEnter the ID of the manual: " + Style.RESET_ALL)
 if not manid:
     os.system('cls' if os.name == 'nt' else 'clear')
     print(Fore.RED + "No ID provided. Please provide a manual id." + Style.RESET_ALL)
@@ -103,12 +112,17 @@ print(Fore.GREEN + f"Total {pn} pages detected." + Style.RESET_ALL)
 os.makedirs(f'out_{manid}', exist_ok=True)
 
 def signal_handler(signal, frame):
+    if get_arg("delete"):
+        print(Fore.RED + "\nKeyboard interrupt detected. No delete flag detected." + Style.RESET_ALL)    
+        print(Fore.RED + "Process interrupted." + Style.RESET_ALL)
+        exit(1)
     print(Fore.RED + "\nKeyboard interrupt detected. Cleaning up temporary files..." + Style.RESET_ALL)
     driver.quit()
     pdf_files = [f'out_{manid}/{file}' for file in os.listdir(f'out_{manid}') if file.endswith('.pdf')]
     for file in pdf_files:
         os.remove(file)
         print(Fore.RED + f"Removed temporary file {file}." + Style.RESET_ALL)
+    os.removedirs(f"out_{manid}")
     os.system('cls' if os.name == 'nt' else 'clear')
     print(Fore.RED + "Process interrupted." + Style.RESET_ALL)
     exit(1)
@@ -120,32 +134,38 @@ if utype == 1:
 else:
     url = f'https://storage.libmanuels.fr/{editor}/manuel/{manid}/1/OEBPS/cover.xhtml?interface=postMessage'
 pdf_file = generate_pdf(driver, url)
-with open(f'out_{manid}/000.pdf', "wb") as outfile:
-    outfile.write(pdf_file.getbuffer())
+if not os.path.exists(f"out_{manid}/000.pdf"):
+    with open(f'out_{manid}/000.pdf', "wb") as outfile:
+        outfile.write(pdf_file.getbuffer())
 print(Fore.BLUE + "Generated PDF for cover." + Style.RESET_ALL)
 for i in range(2, int(pn) + 1):
-    page_number = str(i).zfill(3)
-    if utype == 1:
-        url = f'https://storage.libmanuels.fr/{editor}/specimen/{manid}/1/OEBPS/page{page_number}.xhtml?interface=postMessage'
-    else:
-        url = f'https://storage.libmanuels.fr/{editor}/manuel/{manid}/1/OEBPS/page{page_number}.xhtml?interface=postMessage'
-    pdf_file = generate_pdf(driver, url)
-    with open(f'out_{manid}/{i}.pdf', "wb") as outfile:
-        outfile.write(pdf_file.getbuffer())
+    filename = f'out_{manid}/{i}.pdf'
+    if not os.path.exists(filename):
+        page_number = str(i).zfill(3)
+        if utype == 1:
+            url = f'https://storage.libmanuels.fr/{editor}/specimen/{manid}/1/OEBPS/page{page_number}.xhtml?interface=postMessage'
+        else:
+            url = f'https://storage.libmanuels.fr/{editor}/manuel/{manid}/1/OEBPS/page{page_number}.xhtml?interface=postMessage'
+        pdf_file = generate_pdf(driver, url)
+        with open(filename, "wb") as outfile:
+            outfile.write(pdf_file.getbuffer())
     print(Fore.BLUE + f"Generated PDF for page {i}." + Style.RESET_ALL)
 
 driver.quit()
 
-pdf_files = sorted([f'out_{manid}/{file}' for file in os.listdir(f'out_{manid}') if file.endswith('.pdf')])
+pdf_files = sorted(
+    [f'out_{manid}/{file}' for file in os.listdir(f'out_{manid}') if file.endswith('.pdf')],
+    key=lambda e: int(e.split("/")[1].split(".")[0])
+)
 merge_pdfs(pdf_files, f"manuels/{manid}.pdf")
 print(Fore.MAGENTA + f"Created manual in manuels/{manid}.pdf" + Style.RESET_ALL)
 
 for file in pdf_files:
     os.remove(file)
     print(Fore.RED + f"Removed temporary file {file}." + Style.RESET_ALL)
+os.removedirs(f'out_{manid}')
 
 os.system('cls' if os.name == 'nt' else 'clear')
 print(Fore.GREEN + "Process completed successfully!" + Style.RESET_ALL)
 print(f"It took {Fore.GREEN} %s seconds{Style.RESET_ALL}." % (time.time() - start_time))
-os.removedirs(f'out_{manid}', exist_ok=True)
-os.startfile(f"manuels/{manid}.pdf")
+os.startfile(f"manuels\\{manid}.pdf")
